issue-73-refactor-dashboardpage
