+++ .gitignore
@@ -2,5 +2,6 @@ AI_HANDOFF.md
 server/.env
 client/.env
 LEARNING_NOTES.md
+capstone_defense.md
 server/node_modules/
 client/node_modules/
