+++ client/src/components/dashboard/widgets/CalendarWidget.css
@@ -1,0 +1,136 @@
+.dashboard-widget-calendar .dashboard-widget-content {
+  overflow: hidden;
+  padding-right: 0;
+}
+
+.dashboard-widget-calendar .react-calendar {
+  width: 100%;
+  border: 0;
+  border-radius: 0;
+  background: transparent;
+  color: #345f2f;
+  font-size: 0.66rem;
+  font-family: inherit;
+  line-height: 1.2;
+  padding: 0;
+}
+
+.dashboard-widget-calendar .react-calendar__navigation {
+  margin-bottom: 4px;
+}
+
+.dashboard-widget-calendar .react-calendar__navigation button {
+  min-width: 24px;
+  height: 24px;
+  border: 0;
+  background: transparent;
+  color: #345f2f;
+  font-size: 0.72rem;
+}
+
+.dashboard-widget-calendar .react-calendar__navigation button:hover,
+.dashboard-widget-calendar
+  .react-calendar__navigation
+  button:enabled:focus-visible {
+  background: #ebe4d2;
+}
+
+.dashboard-widget-calendar .react-calendar__month-view__weekdays {
+  text-transform: lowercase;
+  font-size: 0.52rem;
+  color: #5a6956;
+}
+
+.dashboard-widget-calendar .react-calendar__month-view__weekdays__weekday {
+  padding: 3px 0;
+}
+
+.dashboard-widget-calendar .react-calendar__tile {
+  position: relative;
+  height: 24px;
+  padding: 2px 2px 10px;
+  border-radius: 6px;
+  background: transparent;
+  color: #2f3f2c;
+}
+
+.calendar-tile-indicators {
+  position: absolute;
+  left: 50%;
+  bottom: 2px;
+  transform: translateX(-50%);
+  display: inline-flex;
+  align-items: center;
+  gap: 2px;
+}
+
+.calendar-tile-dot {
+  width: 4px;
+  height: 4px;
+  border-radius: 999px;
+}
+
+.calendar-tile-dot-todo {
+  background: #4f7442;
+}
+
+.calendar-tile-dot-bill {
+  background: #c87a22;
+}
+
+.calendar-tile-dot-event {
+  background: #7a5cc9;
+}
+
+.calendar-tile-dot-birthday {
+  background: #d45555;
+}
+
+.calendar-tile-overflow {
+  font-size: 0.46rem;
+  line-height: 1;
+  color: #5a6956;
+  transform: translateY(-1px);
+}
+
+.dashboard-widget-calendar .react-calendar__tile--active .calendar-tile-dot-todo {
+  background: #b9df85;
+}
+
+.dashboard-widget-calendar .react-calendar__tile--active .calendar-tile-dot-bill {
+  background: #f0d279;
+}
+
+.dashboard-widget-calendar .react-calendar__tile--active .calendar-tile-dot-event {
+  background: #c9b7f7;
+}
+
+.dashboard-widget-calendar
+  .react-calendar__tile--active
+  .calendar-tile-dot-birthday {
+  background: #f1bcbc;
+}
+
+.dashboard-widget-calendar .react-calendar__tile--active .calendar-tile-overflow {
+  color: #f8fafc;
+}
+
+.dashboard-widget-calendar .react-calendar__tile:hover,
+.dashboard-widget-calendar .react-calendar__tile:focus-visible {
+  background: #ebe4d2;
+}
+
+.dashboard-widget-calendar .react-calendar__tile--now {
+  background: #e4ecd9;
+  color: #2e5a2b;
+}
+
+.dashboard-widget-calendar .react-calendar__tile--active,
+.dashboard-widget-calendar .react-calendar__tile--active:hover {
+  background: #5b7d4e;
+  color: #ffffff;
+}
+
+.dashboard-widget-calendar .react-calendar__month-view__days__day--neighboringMonth {
+  color: #9aa89a;
+}
