+++ client/src/components/dashboard/widgets/EventsWidget.css
@@ -1,0 +1,65 @@
+.dashboard-widget-events .dashboard-widget-header {
+  padding-right: 2px;
+  margin-bottom: 14px;
+}
+
+.event-list {
+  margin: 0;
+  padding: 0;
+  list-style: none;
+  display: grid;
+  gap: 10px;
+}
+
+.event-item {
+  display: block;
+  font-size: 0.7rem;
+  color: #2f3f2c;
+}
+
+.event-content {
+  min-width: 0;
+  display: grid;
+  row-gap: 3px;
+}
+
+.event-top-row {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 8px;
+  margin-bottom: 0;
+}
+
+.event-title {
+  font-size: 0.72rem;
+  line-height: 1;
+  text-transform: lowercase;
+}
+
+.event-description {
+  margin: 0;
+  color: #566353;
+  font-size: 0.54rem;
+  line-height: 1;
+  font-weight: 500;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.event-meta-row {
+  display: inline-flex;
+  align-items: center;
+  gap: 3px;
+  margin-top: 0;
+}
+
+.event-date,
+.event-time {
+  color: #566353;
+  font-size: 0.54rem;
+  line-height: 1;
+  font-weight: 600;
+  text-transform: lowercase;
+}
