+++ client/src/pages/DashboardPage.css
@@ -93,16 +93,6 @@
   margin-bottom: 8px;
 }
 
-.dashboard-widget-todos .dashboard-widget-header {
-  padding-right: 2px;
-  margin-bottom: 14px;
-}
-
-.dashboard-widget-events .dashboard-widget-header {
-  padding-right: 2px;
-  margin-bottom: 14px;
-}
-
 .dashboard-widget-header .dashboard-widget-title {
   margin: 0;
 }
@@ -133,143 +123,6 @@
 .dashboard-widget-content::-webkit-scrollbar-thumb {
   background: rgba(77, 110, 65, 0.45);
   border-radius: 999px;
-}
-
-.dashboard-widget-calendar .dashboard-widget-content {
-  overflow: hidden;
-  padding-right: 0;
-}
-
-.dashboard-widget-calendar .react-calendar {
-  width: 100%;
-  border: 0;
-  border-radius: 0;
-  background: transparent;
-  color: #345f2f;
-  font-size: 0.66rem;
-  font-family: inherit;
-  line-height: 1.2;
-  padding: 0;
-}
-
-.dashboard-widget-calendar .react-calendar__navigation {
-  margin-bottom: 4px;
-}
-
-.dashboard-widget-calendar .react-calendar__navigation button {
-  min-width: 24px;
-  height: 24px;
-  border: 0;
-  background: transparent;
-  color: #345f2f;
-  font-size: 0.72rem;
-}
-
-.dashboard-widget-calendar .react-calendar__navigation button:hover,
-.dashboard-widget-calendar
-  .react-calendar__navigation
-  button:enabled:focus-visible {
-  background: #ebe4d2;
-}
-
-.dashboard-widget-calendar .react-calendar__month-view__weekdays {
-  text-transform: lowercase;
-  font-size: 0.52rem;
-  color: #5a6956;
-}
-
-.dashboard-widget-calendar .react-calendar__month-view__weekdays__weekday {
-  padding: 3px 0;
-}
-
-.dashboard-widget-calendar .react-calendar__tile {
-  position: relative;
-  height: 24px;
-  padding: 2px 2px 10px;
-  border-radius: 6px;
-  background: transparent;
-  color: #2f3f2c;
-}
-
-.calendar-tile-indicators {
-  position: absolute;
-  left: 50%;
-  bottom: 2px;
-  transform: translateX(-50%);
-  display: inline-flex;
-  align-items: center;
-  gap: 2px;
-}
-
-.calendar-tile-dot {
-  width: 4px;
-  height: 4px;
-  border-radius: 999px;
-}
-
-.calendar-tile-dot-todo {
-  background: #4f7442;
-}
-
-.calendar-tile-dot-bill {
-  background: #c87a22;
-}
-
-.calendar-tile-dot-event {
-  background: #7a5cc9;
-}
-
-.calendar-tile-dot-birthday {
-  background: #d45555;
-}
-
-.calendar-tile-overflow {
-  font-size: 0.46rem;
-  line-height: 1;
-  color: #5a6956;
-  transform: translateY(-1px);
-}
-
-.dashboard-widget-calendar .react-calendar__tile--active .calendar-tile-dot-todo {
-  background: #b9df85;
-}
-
-.dashboard-widget-calendar .react-calendar__tile--active .calendar-tile-dot-bill {
-  background: #f0d279;
-}
-
-.dashboard-widget-calendar .react-calendar__tile--active .calendar-tile-dot-event {
-  background: #c9b7f7;
-}
-
-.dashboard-widget-calendar
-  .react-calendar__tile--active
-  .calendar-tile-dot-birthday {
-  background: #f1bcbc;
-}
-
-.dashboard-widget-calendar .react-calendar__tile--active .calendar-tile-overflow {
-  color: #f8fafc;
-}
-
-.dashboard-widget-calendar .react-calendar__tile:hover,
-.dashboard-widget-calendar .react-calendar__tile:focus-visible {
-  background: #ebe4d2;
-}
-
-.dashboard-widget-calendar .react-calendar__tile--now {
-  background: #e4ecd9;
-  color: #2e5a2b;
-}
-
-.dashboard-widget-calendar .react-calendar__tile--active,
-.dashboard-widget-calendar .react-calendar__tile--active:hover {
-  background: #5b7d4e;
-  color: #ffffff;
-}
-
-.dashboard-widget-calendar .react-calendar__month-view__days__day--neighboringMonth {
-  color: #9aa89a;
 }
 
 .dashboard-error {
@@ -315,63 +168,6 @@
   white-space: nowrap;
 }
 
-.todo-list {
-  margin: 0;
-  padding: 0;
-  list-style: none;
-  display: grid;
-  gap: 10px;
-}
-
-.todo-item {
-  display: block;
-  font-size: 0.7rem;
-  color: #2f3f2c;
-}
-
-.todo-widget-actions {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  gap: 6px;
-  margin-bottom: 0;
-}
-
-.todo-filter-group {
-  display: inline-flex;
-  align-items: center;
-  gap: 8px;
-  height: 16px;
-  margin-right: 6px;
-}
-
-.todo-filter-btn {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  border: 0;
-  background: transparent;
-  color: #345f2f;
-  font-size: 0.54rem;
-  line-height: 1;
-  font-weight: 600;
-  text-transform: lowercase;
-  padding: 0;
-  cursor: pointer;
-}
-
-.todo-filter-btn:hover {
-  text-decoration: underline;
-  text-decoration-thickness: 1.5px;
-  text-underline-offset: 2px;
-}
-
-.todo-filter-btn.is-active {
-  text-decoration: underline;
-  text-decoration-thickness: 2px;
-  text-underline-offset: 2px;
-}
-
 .todo-add-btn,
 .todo-edit-btn {
   display: inline-flex;
@@ -415,187 +211,12 @@
   cursor: not-allowed;
 }
 
-.todo-content {
-  min-width: 0;
-  display: grid;
-  row-gap: 3px;
-}
-
-.todo-top-row {
-  display: flex;
-  gap: 8px;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 0;
-}
-
-.todo-main {
-  display: inline-flex;
-  align-items: center;
-  gap: 8px;
-  min-width: 0;
-  line-height: 1;
-}
-
-.todo-main input[type="checkbox"] {
-  appearance: none;
-  -webkit-appearance: none;
-  margin: 0;
-  width: 12px;
-  height: 12px;
-  border: 1px solid #9fb293;
-  border-radius: 3px;
-  background: #ffffff;
-  display: inline-grid;
-  place-content: center;
-  cursor: pointer;
-}
-
-.todo-main input[type="checkbox"]::before {
-  content: "";
-  width: 6px;
-  height: 6px;
-  border-radius: 2px;
-  background: #4f7442;
-  transform: scale(0);
-  transition: transform 120ms ease-in-out;
-}
-
-.todo-main input[type="checkbox"]:checked::before {
-  transform: scale(1);
-}
-
-.todo-title {
-  font-size: 0.72rem;
-  line-height: 1;
-  text-transform: lowercase;
-}
-
-.todo-description {
-  margin: 0 0 0 20px;
-  color: #566353;
-  font-size: 0.54rem;
-  line-height: 1;
-  font-weight: 500;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
 .todo-row-edit-btn {
   width: 16px;
   height: 16px;
   min-width: 16px;
   padding: 0;
   align-self: center;
-}
-
-.todo-due {
-  color: #566353;
-  font-size: 0.54rem;
-  line-height: 1;
-  font-weight: 600;
-  text-transform: lowercase;
-  white-space: nowrap;
-}
-
-.todo-meta-row {
-  display: inline-flex;
-  align-items: center;
-  gap: 3px;
-  margin-left: 20px;
-  margin-top: 0;
-}
-
-.todo-priority {
-  margin-top: 0;
-  line-height: 1;
-  font-size: 0.54rem;
-  text-transform: lowercase;
-  font-weight: 600;
-}
-
-.todo-item.is-complete .todo-title {
-  text-decoration: line-through;
-  opacity: 0.58;
-}
-
-.todo-item.is-complete .todo-due,
-.todo-item.is-complete .todo-priority {
-  opacity: 0.58;
-}
-
-.priority-high {
-  color: #cc3f3f;
-}
-
-.priority-medium {
-  color: #c87a22;
-}
-
-.priority-low {
-  color: #3e8e52;
-}
-
-.event-list {
-  margin: 0;
-  padding: 0;
-  list-style: none;
-  display: grid;
-  gap: 10px;
-}
-
-.event-item {
-  display: block;
-  font-size: 0.7rem;
-  color: #2f3f2c;
-}
-
-.event-content {
-  min-width: 0;
-  display: grid;
-  row-gap: 3px;
-}
-
-.event-top-row {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 8px;
-  margin-bottom: 0;
-}
-
-.event-title {
-  font-size: 0.72rem;
-  line-height: 1;
-  text-transform: lowercase;
-}
-
-.event-description {
-  margin: 0;
-  color: #566353;
-  font-size: 0.54rem;
-  line-height: 1;
-  font-weight: 500;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.event-meta-row {
-  display: inline-flex;
-  align-items: center;
-  gap: 3px;
-  margin-top: 0;
-}
-
-.event-date,
-.event-time {
-  color: #566353;
-  font-size: 0.54rem;
-  line-height: 1;
-  font-weight: 600;
-  text-transform: lowercase;
 }
 
 .todo-modal-form {
