+++ client/src/pages/DashboardPage.jsx
@@ -1,10 +1,12 @@
-import { useState, useEffect, useMemo, useRef } from "react";
+import { useState, useEffect, useRef } from "react";
 import { Link } from "react-router-dom";
 import { useAuth } from "../auth/AuthContext";
 import "./DashboardPage.css";
 import Modal from "../components/Modal";
-import Calendar from "react-calendar";
-import "react-calendar/dist/Calendar.css";
+import CalendarWidget from "../components/dashboard/widgets/CalendarWidget";
+import TodosWidget from "../components/dashboard/widgets/TodosWidget";
+import EventsWidget from "../components/dashboard/widgets/EventsWidget";
+import PlaceholderWidget from "../components/dashboard/widgets/PlaceholderWidget";
 
 const WIDGETS = ["Calendar", "To-Dos", "Bills", "Events", "Notes", "Birthdays"];
 
@@ -40,7 +42,6 @@ export default function Dashboard() {
   const [isTodoSubmitting, setIsTodoSubmitting] = useState(false);
   const [recentlyCompleted, setRecentlyCompleted] = useState({});
   const recentlyCompletedTimersRef = useRef({});
-  const [calendarDate, setCalendarDate] = useState(new Date());
   const [eventItems, setEventItems] = useState([]);
 
   const name = user?.first_name ?? "there";
@@ -88,25 +89,12 @@ export default function Dashboard() {
   //  Widgets / Dashboard Components
 
   const widgets = dashboardData?.widgets ?? {};
-  const todoCount = widgets.todos?.length ?? 0;
   const notesCount = widgets.notes?.length ?? 0;
   const billsCount = widgets.bills?.length ?? 0;
-  const eventsCount = widgets.events?.length ?? 0;
   const birthdaysCount = widgets.birthdays?.length ?? 0;
-  const calendarCount = widgets.calendar?.length ?? 0;
 
   const getWidgetMessage = (name) => {
     switch (name) {
-      case "Calendar":
-        return calendarCount === 0
-          ? "No calendar items yet."
-          : `${calendarCount} item(s) loaded.`;
-
-      case "To-Dos":
-        return todoCount === 0
-          ? "No to-dos yet."
-          : `${todoCount} item(s) loaded.`;
-
       case "Notes":
         return notesCount === 0
           ? "No notes yet."
@@ -116,11 +104,6 @@ export default function Dashboard() {
         return billsCount === 0
           ? "No bills yet."
           : `${billsCount} item(s) loaded.`;
-
-      case "Events":
-        return eventsCount === 0
-          ? "No events yet."
-          : `${eventsCount} item(s) loaded.`;
 
       case "Birthdays":
         return birthdaysCount === 0
@@ -505,63 +488,6 @@ export default function Dashboard() {
     });
   };
 
-  const toDateKey = (value) => {
-    const d = parseDateValue(value);
-    if (!d) return null;
-    const y = d.getFullYear();
-    const m = String(d.getMonth() + 1).padStart(2, "0");
-    const day = String(d.getDate()).padStart(2, "0");
-    return `${y}-${m}-${day}`;
-  };
-
-  const calendarIndicatorsByDate = useMemo(() => {
-    const map = {};
-
-    const add = (dateValue, type) => {
-      const key = toDateKey(dateValue);
-      if (!key) return;
-      if (!map[key]) map[key] = [];
-      map[key].push(type);
-    };
-
-    todoItems.forEach((t) => add(t.due_date, "todo"));
-    (widgets.bills ?? []).forEach((b) =>
-      add(b.due_date ?? b.next_due_date, "bill"),
-    );
-    const eventSource = eventItems.length > 0 ? eventItems : (widgets.events ?? []);
-    eventSource.forEach((e) => add(e.event_date ?? e.date, "event"));
-    (widgets.birthdays ?? []).forEach((b) =>
-      add(b.birthday ?? b.birth_date, "birthday"),
-    );
-
-    return map;
-  }, [todoItems, eventItems, widgets.bills, widgets.events, widgets.birthdays]);
-
-  const renderCalendarTileContent = ({ date, view }) => {
-    if (view !== "month") return null;
-
-    const key = toDateKey(date);
-    const indicators = calendarIndicatorsByDate[key] ?? [];
-    if (indicators.length === 0) return null;
-
-    const visible = indicators.slice(0, 3);
-    const overflow = indicators.length - visible.length;
-
-    return (
-      <div className="calendar-tile-indicators" aria-hidden="true">
-        {visible.map((type, idx) => (
-          <span
-            key={`${type}-${idx}`}
-            className={`calendar-tile-dot calendar-tile-dot-${type}`}
-          />
-        ))}
-        {overflow > 0 ? (
-          <span className="calendar-tile-overflow">...</span>
-        ) : null}
-      </div>
-    );
-  };
-
   // EVENTS WIDGET //
 
   useEffect(() => {
@@ -632,222 +558,65 @@ export default function Dashboard() {
           </p>
         )}
         <section className="dashboard-grid" aria-label="Dashboard widgets">
-          {WIDGETS.map((name) => (
-            <article
-              key={name}
-              className={`dashboard-widget dashboard-widget-${name
-                .toLowerCase()
-                .replace(/[^a-z]/g, "")}`}
-            >
-              <div className="dashboard-widget-header">
-                <h2 className="dashboard-widget-title">{name}</h2>
-                {name === "To-Dos" && (
-                  <div className="todo-widget-actions">
-                    <div className="todo-filter-group">
-                      <button
-                        type="button"
-                        className={`todo-filter-btn ${todoFilter === "today" ? "is-active" : ""}`}
-                        onClick={() => setTodoFilter("today")}
-                      >
-                        today
-                      </button>
-                      <button
-                        type="button"
-                        className={`todo-filter-btn ${todoFilter === "week" ? "is-active" : ""}`}
-                        onClick={() => setTodoFilter("week")}
-                      >
-                        week
-                      </button>
-                      <button
-                        type="button"
-                        className={`todo-filter-btn ${todoFilter === "all" ? "is-active" : ""}`}
-                        onClick={() => setTodoFilter("all")}
-                      >
-                        all
-                      </button>
-                      <button
-                        type="button"
-                        className={`todo-filter-btn ${todoFilter === "completed" ? "is-active" : ""}`}
-                        onClick={() => setTodoFilter("completed")}
-                      >
-                        completed
-                      </button>
-                    </div>
-                    <button
-                      type="button"
-                      className="todo-add-btn"
-                      onClick={() => openCreateModal("To-Dos")}
-                    >
-                      +
-                    </button>
+          {WIDGETS.map((name) => {
+            if (name === "To-Dos") {
+              return (
+                <TodosWidget
+                  key={name}
+                  todoFilter={todoFilter}
+                  setTodoFilter={setTodoFilter}
+                  visibleTodos={visibleTodos}
+                  toggleTodoCompleted={toggleTodoCompleted}
+                  openCreateModal={openCreateModal}
+                  openEditModal={openEditModal}
+                  formatDueDate={formatDueDate}
+                />
+              );
+            }
+
+            if (name === "Events") {
+              return (
+                <EventsWidget
+                  key={name}
+                  events={events}
+                  openCreateModal={openCreateModal}
+                  openEditModal={openEditModal}
+                  formatEventDate={formatEventDate}
+                  formatEventTime={formatEventTime}
+                />
+              );
+            }
+
+            if (name === "Calendar") {
+              return (
+                <article
+                  key={name}
+                  className="dashboard-widget dashboard-widget-calendar"
+                >
+                  <div className="dashboard-widget-header">
+                    <h2 className="dashboard-widget-title">Calendar</h2>
                   </div>
-                )}
-                {name === "Events" && (
-                  <button
-                    type="button"
-                    className="todo-add-btn"
-                    onClick={() => openCreateModal("Events")}
-                  >
-                    +
-                  </button>
-                )}
-              </div>
-              <div className="dashboard-widget-content">
-                {name === "Calendar" ? (
-                  <Calendar
-                    onChange={setCalendarDate}
-                    value={calendarDate}
-                    tileContent={renderCalendarTileContent}
-                  />
-                ) : name === "To-Dos" ? (
-                  visibleTodos.length === 0 ? (
-                    <p className="dashboard-widget-placeholder">
-                      {todoFilter === "completed"
-                        ? "No completed to-dos yet."
-                        : "No to-dos yet."}
-                    </p>
-                  ) : (
-                    <ul className="todo-list">
-                      {visibleTodos.map((todo) => {
-                        const hasDescription = Boolean(
-                          todo.description?.trim(),
-                        );
-                        return (
-                          <li
-                            key={todo.id}
-                            className={`todo-item ${todo.completed ? "is-complete" : ""}`}
-                          >
-                            <div className="todo-content">
-                              <div className="todo-top-row">
-                                <label className="todo-main">
-                                  <input
-                                    type="checkbox"
-                                    checked={Boolean(todo.completed)}
-                                    onChange={() => toggleTodoCompleted(todo)}
-                                    aria-label={`Mark ${todo.title} complete`}
-                                  />
-                                  <strong className="todo-title">
-                                    {todo.title}
-                                  </strong>
-                                </label>
-                                <button
-                                  type="button"
-                                  className="todo-edit-btn todo-row-edit-btn"
-                                  onClick={() => openEditModal("To-Dos", todo)}
-                                  aria-label={`Edit ${todo.title}`}
-                                  title={`Edit ${todo.title}`}
-                                >
-                                  <svg
-                                    className="todo-edit-icon"
-                                    viewBox="0 0 24 24"
-                                    aria-hidden="true"
-                                  >
-                                    <path
-                                      d="M16.862 3.487a2.25 2.25 0 1 1 3.182 3.182L9.75 16.963 6 18l1.037-3.75 9.825-9.825Z"
-                                      fill="currentColor"
-                                    />
-                                    <path
-                                      d="M19.5 13.5v5.25A2.25 2.25 0 0 1 17.25 21H5.25A2.25 2.25 0 0 1 3 18.75V6.75A2.25 2.25 0 0 1 5.25 4.5H10.5"
-                                      fill="none"
-                                      stroke="currentColor"
-                                      strokeWidth="1.5"
-                                      strokeLinecap="round"
-                                    />
-                                  </svg>
-                                </button>
-                              </div>
-                              {hasDescription ? (
-                                <p className="todo-description">
-                                  {todo.description.trim()}
-                                </p>
-                              ) : null}
-                              <div className="todo-meta-row">
-                                <span
-                                  className={`todo-priority priority-${todo.priority}`}
-                                >
-                                  {todo.priority}
-                                </span>
-                                <span className="todo-due">
-                                  {formatDueDate(todo.due_date)}
-                                </span>
-                              </div>
-                            </div>
-                          </li>
-                        );
-                      })}
-                    </ul>
-                  )
-                ) : name === "Events" ? (
-                  events.length === 0 ? (
-                    <p className="dashboard-widget-placeholder">
-                      No events yet.
-                    </p>
-                  ) : (
-                    <ul className="event-list">
-                      {events.map((event) => (
-                        <li key={event.id} className="event-item">
-                          <div className="event-content">
-                            <div className="event-top-row">
-                              <strong className="event-title">
-                                {event.title}
-                              </strong>
-                              <button
-                                type="button"
-                                className="todo-edit-btn todo-row-edit-btn"
-                                onClick={() => openEditModal("Events", event)}
-                                aria-label={`Edit ${event.title}`}
-                                title={`Edit ${event.title}`}
-                              >
-                                <svg
-                                  className="todo-edit-icon"
-                                  viewBox="0 0 24 24"
-                                  aria-hidden="true"
-                                >
-                                  <path
-                                    d="M16.862 3.487a2.25 2.25 0 1 1 3.182 3.182L9.75 16.963 6 18l1.037-3.75 9.825-9.825Z"
-                                    fill="currentColor"
-                                  />
-                                  <path
-                                    d="M19.5 13.5v5.25A2.25 2.25 0 0 1 17.25 21H5.25A2.25 2.25 0 0 1 3 18.75V6.75A2.25 2.25 0 0 1 5.25 4.5H10.5"
-                                    fill="none"
-                                    stroke="currentColor"
-                                    strokeWidth="1.5"
-                                    strokeLinecap="round"
-                                  />
-                                </svg>
-                              </button>
-                            </div>
-                            {event.description?.trim() ? (
-                              <p className="event-description">
-                                {event.description.trim()}
-                              </p>
-                            ) : null}
-                            <div className="event-meta-row">
-                              <span className="event-date">
-                                {formatEventDate(event.event_date)}
-                              </span>
-                              {event.start_time ? (
-                                <span className="event-time">
-                                  {formatEventTime(event.start_time)}
-                                  {event.end_time
-                                    ? ` - ${formatEventTime(event.end_time)}`
-                                    : ""}
-                                </span>
-                              ) : null}
-                            </div>
-                          </div>
-                        </li>
-                      ))}
-                    </ul>
-                  )
-                ) : (
-                  <p className="dashboard-widget-placeholder">
-                    {getWidgetMessage(name)}
-                  </p>
-                )}
-              </div>
-            </article>
-          ))}
+                  <div className="dashboard-widget-content">
+                    <CalendarWidget
+                      todoItems={todoItems}
+                      eventItems={eventItems}
+                      dashboardEvents={widgets.events}
+                      bills={widgets.bills}
+                      birthdays={widgets.birthdays}
+                    />
+                  </div>
+                </article>
+              );
+            }
+
+            return (
+              <PlaceholderWidget
+                key={name}
+                name={name}
+                message={getWidgetMessage(name)}
+              />
+            );
+          })}
         </section>
       </section>
       <Modal
